better-archive-creation
Create an archive that contains the _directory_ containing the files, not an archive that contains the files with no surrounding container. 
Doing it this way ensures that when you run `tar -xzf ${downloadDirectoryArchive}.tgz` you will get a _directory_, and _that_ will contain the files. 
Otherwise, you will end up filling your current working directory with the contents of the archive.